refactor(files): relocate waitForPendingFiles helper to services/files
The helper polls file-processing state across both queue and SSE chat
paths, so the services/queue/workers namespace no longer fits.

Move src/services/queue/workers/fileWait.ts to
src/services/files/waitForPending.ts (and its tests alongside). Update
the only static caller (chat.worker.ts dynamic import). Helper types
stay imported from services/queue/types — the Queue<> parameter is part
of the helper's contract, not an accident of its old location.

Existing unit tests at __tests__/waitForPending.test.ts move with the
file and continue to characterize the helper's behavior unchanged.
